improve forgotten password page
